Refactor UpgradeTier to separate file
